[includes] Enforce IWYU on tests/
This also includes some fixes of the Bazel rules for tests.

When building for IWYU, we need a special flag to build the CPU
tests as well (since they are disabled if a GPU backend is built
by default).

This does not yet add a CI job that actually compiles the tests,
but that is coming soon.

Change-Id: Ie7f38a932a0b3112f0d3915640924b6b48925f30
Bug: skia:13052
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/592538
Reviewed-by: Ben Wagner <bungeman@google.com>