Fix fuzzer-discovered error with interface blocks.
The fuzzer discovered that interface blocks can redefine a built-in
variable like `sk_Clockwise`. When this happens, we would look up
`sk_Clockwise` and expect to find a Variable, but it's actually a
Field, so we would assert.

Now, we will check that the symbol is a Variable and stop processing
it if it's not one. This doesn't fix the root issue--InterfaceBlocks
should not be allowed to pave over built-ins at all. I've filed a
separate bug (skia:13823) to track this issue.

Change-Id: I5b1d16d08d5005f543e15a8b0de1f10dc1c4a950
Bug: oss-fuzz:52509, skia:13823
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/592741
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>