Reject global symbols that overlap module globals.
Generally, SymbolTables have behaved as if their layering matches the
layering of the program; i.e., the top level SymbolTable is the top
level of the program, and nested SymbolTables correspond to scopes in
the code. This is usually true, but modules break that assumption; all
modules add things at global scope, and then the program _also_ adds
things on top of the module but also at global scope.

Up until now, we didn't properly account for this when looking for
duplicate symbol names; SkSL would silently allow program globals
which shadowed module globals.

Now, if we are adding a symbol at the symbol boundary, we also look
in the modules for a matching symbol, and if one is found, we reject
the addition.

We actually relied on this ability in the ModuleLoader, where we
inject invalid symbols to disable features in certain ProgramKinds.
To account for this, there is now a SymbolTable API `inject` which
will force in a symbol, regardless of the presence of any other
symbol with the same name. The ModuleLoader now uses this API.

Change-Id: I9917ddaf5d0b2abe1ad22c6d693603d679337118
Bug: oss-fuzz:52509, skia:13823
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/592959
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>