[graphite] Add vertex shader to AnalyticRRectRenderStep
This positions vertices for filled rects, round rects, and stroked
rects and round rects with bevel,miter, and round corners, and
hairlines. The vertex shader is also implemented to work with general
convex quadrilaterals, but there's no way to record them in the
DrawList and there's no fallback logic for triangles, lines, or points
(quads with one or two duplicates).

This does not yet implement the coverage calculations so rectangles
are aliased and 1px wider, and round rects are octogons.

Bug: b/241402913
Bug: b/241403207
Bug: b/241403109
Bug: b/241403427
Change-Id: If9ddc976c9993c82df1bdbb01d7b4cc5fc3fbf75
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/578498
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>