Return a DSLExpression from InterfaceBlock, not a DSLGlobalVar.
When we create an InterfaceBlock, it creates a global variable
associated with the block. If it's an anonymous interface block,
the global has no name, but it's still necessary under the hood--
the individual interface block field objects are Field
expressions, not Variable expressions, and Fields need an owner.

Previously, we returned a DSLGlobalVar from InterfaceBlock, which
was the global variable itself. However, the caller doesn't want
the variable really--it wants a VariableReference to it, so it
can say `myVarRef.field("whatever")`. Previously this only worked
due to the silent conversion from variable to expression in DSL.

(Note that the real user of this code, the Parser, doesn't care
anyway--it ignores the result entirely.)

Change-Id: Ie7af51434203db4c73b0d2d59a08dfd8e7896f06
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/593425
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>