[Graphite] Don't store numLevels but just Mipmapped on TextureInfo.
TextureInfo was meant to be independent of the size of a Texture. But
by have the actual numMipLevels on it, it indirectly was dependent on
knowing the size. Additionally we didn't actually use the number of
levels except to either know we were mipped or to verify the correct
number of levels were being used for an upload.

Change-Id: I5db11daee6b87478ee03049c83e2d7cd8e21d62e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/593963
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>