Truncate very long error reporting lines.
The fuzzer has discovered that it can consume massive amounts of
memory by creating very long lines that contain lots of errors. We
echo the line to the error text once per error, so if the line is
thousands of characters long with hundreds of errors, this can be
a significant amount of text.

The SkMeshSpecification fuzzer makes it easy to generate large
amounts of text, because (in order to give the fuzzer a better
chance of success) we directly translate certain input bytes into
the names of our attributes or varyings. In this case, the fuzzer
has made a very very long attribute/varying name and used the
control byte to repeat it over and over again in the input, then
then synthesized a ton of errors on the same line. The fuzzed
program in this case generated 354 errors and the error log was
over 900MB of text in total.

To work around this issue, the error reporter now only shows a
maximum of 100 characters before or after any error marks. For
typical human-authored code, this will generally show the entire
line, but for minified or fuzzed code, this dramatically improves
the readability of reporting, and limits memory consumption.

Change-Id: I01d22d6c16919e00a4aad8ecc018e0f1b120b5f2
Bug: oss-fuzz:52568
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/594219
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>