Detect Mesh local coordinates that are just a varying passthrough.
When changing mesh user shader signatures we lost the ability to
lift local matrix application to the vertex shader for the case of
position == local coords. Because drawVertices is implemented on top
of SkMeshSpecification this applies to vertex draws as well.

Now we detect whether the user's FS shader simply returns *any*
varying as the local coords and lift local matrix application to the
vertex shader. This approach also catches the case of SkVertices with
separate texture coordinates.

The current implementation is limited to cases of simple
"return varyings.foo;". Multiple returns are detected if they are
all the same varying. Swizzles defeat the optimization.

Bug: skia:12720
Change-Id: I42cca3d8d5017c699157866c6b2dcaffd85f3f9f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/593424
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Brian Salomon <bsalomon@google.com>