Use scaled blits for backdrop filters
SaveLayerRecs created for ChromeOS's backdrop blurs were already
specifying a downscale factor that was getting backed into the device/
intermediate transforms of the layer's skif::Mapping. This was always
going through a draw, which means when the backdrop was the main
framebuffer we'd do a full framebuffer copy and then downscale, and
then blur.

This adds a new snapSpecialScaled() API to SkDevice() so that backends
that support scaling while copying/snapping (e.g. Ganesh using
rescaleInto()), and then updates the logic in
SkCanvas::internalDrawDeviceWithFilter() to attempt scaling with the
snap before falling back to its draw.

This should be the last change needed to bring SkiaRenderer's backdrop
filtering performance back to what had been possible in GlRenderer
on ChromeOS (which always used glBlitFramebuffer to downscale the
main screen).

Bug: b/197774543
Change-Id: I8c50d3170037f66a50dab35eeee4f5762993cd5f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/582500
Reviewed-by: Brian Salomon <bsalomon@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>