Implement gradient color space options in SkVM
We'll want an SkRasterPipeline version eventually, but this was
straightforward, and ensures that the new features always have a
working fallback (eg, if the paint also includes SkSL).

This is a direct implementation of the conversion functions from the
CSS specification. We'll be optimizing all of these to remove
unnecessary steps soon.

Bug: skia:13108
Change-Id: Ib70ec5bc2a5c12445a4a16c676bf2075b68f66f1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/595757
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>