[bazel] Get CPU tests to pass on RBE.
bazel test //tests/... --config=cpu_only_debug --config=linux_rbe \
    --remote_download_minimal --test_output=errors

All but one of the tests passed initially. The failing test was one
that required a default font to have Bold, Italic, and Bold+Italic
variants. [1]

As it turns out, our RBE image only had a few fonts from the
DejaVu family, but no italic variants.

Thus, this CL updates the RBE image to include not only fonts
to make that test pass, but also fonts needed for GMs (when
we need to run those). However, we needed to go to an older
version of Debian because the latest Debian 12 (Bookworm)
version has a version of GLIBC that causes an issue with the
version of Docker used by RBE (20.10.8) [2].

Using an older RBE image required that the version of Clang
we use could be run with an older version of GLIBC. The Clang
15 binary available from GitHub [3] requires GLIBC 2.33, but
Debian 11 (Bullseye) only has 2.31.

Thus, in a previous CL [4], we changed Bazel to use the Clang
which we compile ourselves in a Debian 10 (buster) image for
use with the GN build.

With the GLIBC, Clang, and Docker situation sorted out, we
could add the fonts to the RBE image, following the steps
in //bazel/rbe/README.md.

The test still did not pass because none of the default font
family names that src/ports/SkFontMgr_custom.cpp tries to use
were installed (many of which are proprietary). So, we updated
that list to find the DejaVu Serif font, which is on many Linux
distros by default, and has Italic, Bold, and BoldItalic variants.

For good measure, we update the version of Bazel to a stable version
and not a release candidate.

A follow-on CL will add a CI job to run the tests.

Suggested Review Order:
 - src/ports/SkFontMgr_custom.cpp
 - bazel/rbe/gce_linux_container/Dockerfile to see the changes to
   the image.
 - bazel/platform/BUILD.bazel to see where the image hash needs
   to be updated such that it is used.
 - bazel/rbe/README.md
 - IGNORE everything in bazel/rbe/gce_linux/... as that is all
   generated
 - Everything else.

[1] https://github.com/google/skia/blob/0c437ea3502af8407c9f13fa99d4a2d84830a765/tests/TypefaceTest.cpp#L599-L611
[2] https://medium.com/nttlabs/ubuntu-21-10-and-fedora-35-do-not-work-on-docker-20-10-9-1cd439d9921
[3] https://github.com/llvm/llvm-project/releases/download/llvmorg-15.0.2/clang+llvm-15.0.2-x86_64-unknown-linux-gnu-rhel86.tar.xz
[4] https://skia-review.googlesource.com/c/skia/+/594807

Change-Id: I1b52928fe2922745387a4b2e31c98cb4862a4d41
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/593557
Reviewed-by: Ben Wagner <bungeman@google.com>