[bazel] Update Mac toolchain to use Clang 15.0.1
15.0.1 was chosen because this is the same version we build for
Linux [1] and I would like to keep all OSes on the same version.

This adds a arm64 specific version of the compiler that is downloaded
if the host machine is an M1 mac. Even though the x86_64 version
can be used on my M1 mac, the amd64 compiler was ~1.5 times faster
(25s with arm64 version, 37s with amd64 version)

If we ever do remote Mac builds, this might need to change to
have both arm64 and amd64 as separate toolchain filegroups so
the executor Mac gets the right version for its chipset, but this
should work for the local builds we do currently.

[1] https://skia-review.googlesource.com/c/skia/+/594807

Change-Id: Ie70e0e43c6d0f077f4c9a49dcd7edcaf790bc024
Bug: skia:12541
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/593423
Reviewed-by: Jorge Betancourt <jmbetancourt@google.com>