Improve error handling with unknown types.
Previously, if we were attempting to parse a type and found an
identifier (a word) that wasn't in the symbol table, we returned
a null type, which is interpreted as "this is not a type at all" by
the caller. We now return the invalid type, which tells the caller
"yes, this was a type, but an invalid one" and allows the parse to
continue as if a type was found in that spot. This gives better
parsing results for a misspelled type name.

Change-Id: I12a6a129358a7f6c9f504abcbc1ac0b79491553f
Bug: skia:13383
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/598084
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>