[bazel] Add rules to run SkSL Minify
With the GN build, one could generate a GN build with
  skia_compile_modules=true
  skia_compile_sksl_tests=true
and then run
  ninja -C out/Generate minify_sksl
  ninja -C out/Generate minify_sksl_tests
in order to regenerate .sksl files (and more) in
//tests/sksl/* using source code in //resources/sksl/*

The Bazel equivalent for these commands is now:
  bazel run //tools/sksl-minify:minify_srcs
  bazel run //tools/sksl-minify:minify_tests

Because we want to check in these generated files (that is,
they live in the source repo), we need to make a py_binary
that can run from the Skia root (and not deep in the Bazel
exec folder), so it can overwrite the checked in minified
files if necessary. We want this to be a Python script
(which runs with Bazel's hermetic Python runtime) to be
cross-platform and easier to maintain than Bash/Powershell.
See the genrules in tools/sksl-minify/BUILD.bazel for this.

The GN version calls out to gn/minify_sksl.py and
gn/minify_sksl_tests.py. The former is really simple, and
I had the Bazel equivalent skip it. The latter is more
involved, so I have Bazel call it (also using the hermetic
Python runtime).

Known issue:
 - sksl_minify won't bootstrap (i.e. run in standalone mode)
   until the skslc change lands.

Suggested Review Order:
 - tools/sksl-minify/BUILD.bazel, noting that we compile
   SkSLMinify.cpp and use it from the generated Python script.
   For generating tests, we emulate the existing behavior
   of the GN version and create a file with inputs to avoid
   exceeding the command line limit.
 - src/sksl/BUILD.bazel and resources/sksl/BUILD.bazel to
   see the files which need to be minified.
 - All other files.

Change-Id: I826c9e7ca8780a3a1e61f98d656543dd7ab44889
Bug: b/256869968
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/599856
Reviewed-by: Ben Wagner <bungeman@google.com>