Fix case when a non-mipmapped draw adds dependency before mipped draw
Previously if a render task was adding a dependency to a proxy and the
task was already dependent on the proxy's last render task it would
skip adding the proxy to a resolve task.

In release builds this could lead to drawing using stale mip level content. In debug it would trip an assert that the texture's mipmaps were clean.

Bug: chromium:1355807
Bug: chromium:1365578

Change-Id: If68e2c52f8ae532ffee458d5d8bcd3319ede4399
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/600396
Commit-Queue: Brian Salomon <bsalomon@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>