[sksl] Modify sksl_compile_tests to be more Bazel friendly
Changes the inputs to be a list of resources, like
http://review.skia.org/599346/

This changes where SkSLModuleLoader looks for .sksl files in
standalone mode. It now looks in the same folder as the
executable instead of the current working directory. This
is a lot easier to work with in Bazel and should not be
a change for the GN build.

Change-Id: I9cf19ba2d55191d45a09f80f10a7903174492cef
Bug: b/256870330
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/600097
Reviewed-by: Ben Wagner <bungeman@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>