[bazel] Add rules to compile SkSL tests
Similar to http://review.skia.org/599856, this adds a Bazel rule
to run skslc (the SkSL Compiler) on a bunch of sksl inputs
(and more) to generate test outputs in //tests/sksl/. Because
there are many variants of the rule, we use a macro called
compile_sksl to cut down on the boilerplate (similar to [1]).

This backports a change to expose skia_cc_binary_with_flags
as a macro, and use it here and for sksl-minify.

These can all be run with a convenient Make target:
    make -C tools/skslc compile_all_tests

Suggested review order:
 - tools/skslc/compile_sksl.bzl to see the macro which generates
   a py_binary that we run and a genrule which creates a list
   of input files.
 - tools/skslc/BUILD.bazel which uses the macro repeatedly
   porting the existing GN rules [1].
 - resources/sksl/BUILD.bazel which emulates sksl_tests.gni.
   I hope the globs are short-lived and soon we have the Bazel
   file be the source of truth with the .gni files generated
   from the Bazel file.
 - All other changes.

[1] https://github.com/google/skia/blob/1272b520c082ed0b681bd10b9ff5d7663a1670fb/BUILD.gn#L804-L891
[2] https://github.com/google/skia/blob/1272b520c082ed0b681bd10b9ff5d7663a1670fb/gn/sksl_tests.gni

Change-Id: Iad9873c51ce4486f7e75d7b3b2a79201e2f7fcba
Bug: b/256870330
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/600276
Reviewed-by: Ben Wagner <bungeman@google.com>