Viewer slides default to window-filling dimensions
The Slide abstraction has a size that is independent of the window size
that is intended the be the size of the content the slide displays.
However, this isn't always a meaningful distinction as many slide
subclasses just fill the window with content. Make that the default
behavior so getDimensions() only needs to be overridden for subclasses
that have a separate notion of content size.

Bug: skia:13870
Change-Id: Ie11fe47c0dc325cebfcdd879a1d3811bec075041
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/601318
Commit-Queue: Brian Salomon <bsalomon@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>