[bazel] Reorganize defines and local_defines to defines.bzl
The primary motivation for this is fixing when SKIA_IMPLEMENTATION
is set.

We only want to set SKIA_IMPLEMENTATION=1 (and thereby configure
include/private/gpu/vk/SkiaVulkan.h to include our private version
of Vulkan) when we are compiling Skia .cpp files, but not when
clients are using Skia (so they can use whatever version of Vulkan
they like).

This reorganizes the defines to be in a defines.bzl file
(which can get synced to G3). This allows us to break up the
defines into "public" defines and local_defines.

See also http://cl/487001769

Change-Id: I85203bafeba0b760adc1a651f725953b25bfdefa
Bug: skia:12541
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/601623
Reviewed-by: Greg Daniel <egdaniel@google.com>
Reviewed-by: Ben Wagner <bungeman@google.com>