Convert FatBits and ManyRects from Samples to Slides
I don't think we actually need the Sample abstraction. It's a legacy
from the old SampleApp. The largest porting difference is how clicks
are handled. Add a bridge class that reimplements the Sample Click
interface as a Slide.

Also add a helper for slides that don't have a built in content size
but instead fill the window, whatever its size.

Bug: skia:13870
Change-Id: I322075ae0e5e72621e6bb139ed088c41329042f2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/598086
Commit-Queue: Brian Salomon <bsalomon@google.com>
Reviewed-by: Herb Derby <herb@google.com>