[experiment] Replace ball of void* with SkRasterPipelineStage struct.
This might be rolled back depending on results from perf.

This replaces two lodsq instructions with 1-2 movq insteads and
an addq for every stage. This costs us a fair amount of code size
(4 bytes per stage -> 8 or 12 bytes per stage) but the movq/addq
instructions are more efficient on most hardware, so it might be a
win.

Change-Id: I8510719f32eff38a48c1330d6e5938f78cc07968
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/601598
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>