[graphite] Add primitive color blending and apply paint alpha value
Fix primitive blending functionality, enabling legible viewer application toolbar view.

To observe this work as intended in the viewer application, launch the viewer application on the Vertices slide and press the spacebar. You should see legible text which demonstrates primitive blending working as intended. The window is jarringly red which is likely due to graphite's texture sampling not yet conveying the SkColorType correctly; as in, not yet modifying a `sample` call return value with the appropriate swizzle.

Change-Id: I7c8e94706539f1df4b77c45b3183e7fa11c6a4b6
Bug: b/246583457
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/580417
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Nicolette Prevost <nicolettep@google.com>