[sksl][spirv] Skip OpVariable declarations only for constant scalars and vectors
Other composite types may be accessed with a dynamic index variable
which requires the use of OpAccessChain, which must operate over a base
operand of pointer variable.

The logic is now changed to always generate an OpVariable for all constant
variables with a compile-time constant value except for scalars and
vectors.

Alternatly we could try to optimize this and defer a variable declaration
until the constant composite is actually accessed with a dynamic index
but that doesn't seem to be worth it.

Bug: oss-fuzz:53202
Change-Id: I345a974bd7c969c82caa4e8d7b26f8782ca5a5d4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/603405
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>