Move data pointer for move operator=
The original implementation just moved each element over to
the destination Array. It maybe that it was discovered that
moving the pointer didn't pay off. I don't know if this was
tested or not.

Move the pointer from that to this if possible. Otherwise,
move the elements one-by-one.

Change-Id: I8c1dae1a79e46bb230c8eb84be723e0c05ba7459
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/605718
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Herb Derby <herb@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>