[bazel] Support enforcing IWYU on select headers
We don't compile C++ headers on their own, only .cpp files.
As a result, IWYU is not run on headers which do not have
a .cpp file.

To remedy this, we can generate an .cpp file that is the
contents of a header file that we want to enforce IWYU
on. The generation happens *inside* of Bazel's compilation
directory, so it is not checked in nor in the source tree.

This shows an example for enforcing IWYU on SkRefCnt.h.
In the include/BUILD.bazel file,
//include/gen/include/core/SkRefCnt.h.cpp is
generated and included into the Skia build. This allows
IWYU to notice extraneous or missing headers, which
the rest of the CL takes care of.

Suggested Review Order
 - bazel/generate_cpp_files_for_headers.bzl
 - include/BUILD.bazel
 - bazel/copts.bzl to ignore warnings that only happen when
   compiling a header as if it were a source file.
 - IWYU_mapping.imp (for many fixes where symbols were
   "public" when the docs say they should be "private")
 - All other changes, which were all mechanically done.

Change-Id: I791e011714841bcbfc96555ce9fc63806cdde170
Bug: skia:13052
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/603402
Reviewed-by: Herb Derby <herb@google.com>
Reviewed-by: Ben Wagner <bungeman@google.com>
Owners-Override: Kevin Lubick <kjlubick@google.com>