Add a stack for condition masks.
The builder keeps track of the stack position and max-depth for us;
push and pop commands turn into stores and loads at the correct index.
(This approach limits our code structure a bit--for instance, this
only makes sense if functions are fully inlined. We could burn a
register on a stack index if this were a significant issue, but it
probably is OK as-is.)

Change-Id: I6ea2f4e88d1ec0b2b2e60cb4dc101c5a4f5ae83b
Bug: skia:13676
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/607496
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>