Fix skvm gradient interval calculation
By clamping t first, we prevented ourselves from ever using the first
(fake) interval, even when t was negative (in the clamp region). The
result was that we evaluated the second interval using a negative t,
effectively extrapolating backwards past the first color, rather than
just using the interval we had already computed to produce that color
exactly.

Instead, keep t unclamped, then clamp the index result to the size of
the interval table before the lookup. This bug was really visible on the
new gradients_color_space_many_stops GM, but fixing it has shown that it
was also present on many of the COLRv1 GMs, and was (slightly) visible
on some of the very basic gradient GMs (which now look much more like RP
and GPU).

Change-Id: I24e4810b8ea5dde7875b962046f07f773fc391ad
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/608179
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>