Add temp value stack to RasterPipelineBuilder.
At present the only thing you can do with the temp-value stack is
push immediates onto it, and discard values from it. It's not a very
useful stack yet, but there's enough code here to peel off a CL for
review.

Like the condition-mask stack, the Builder is responsible for
reserving the right amount of space for the stack based on the
program. It is also responsible for tracking the stack position
across pushes and pops. This CL doesn't actually introduce any new
SkRasterPipeline ops, because push-immediate can be represented with
existing ops (immediate_f, store_unmasked).

Change-Id: Ib5d8faa1ab5b32ad978c5571fd39ce183b4f0822
Bug: skia:13676
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/608021
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>