Add copy-slots-unmasked to RP.
This is structurally almost the same as copy-slots-masked, but the
actual op itself is just a memcpy. This should be faster, so we can
use it in cases where ignoring the mask is okay. For instance,
when pushing a value onto the temp stack, copying every lane is
fine. It's only when we copy back into the value slot that the mask
is important.

Change-Id: If72ee4bbea76bbb79b759029ffa9eb51bed62fcb
Bug: skia:13676
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/608498
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>