[graphite][compute] Use BindBufferInfo for ComputePassTask buffer bindings
Instead of taking shared ownership of buffer objects, a ComputePassTask
is now initialized with BindBufferInfos. This allows it to use
pooled and externally managed buffers (such as those in
DrawBufferManager).

As we extend the ResourceBinding struct to support different resource
types (i.e. Texture/Sampler objects) we can eventually introduce an
"owned buffer" type if we need to a compute task to track a
sk_sp<Buffer>.

Change-Id: I72d3266d1aa337d7279720b5fdd92bd442503bea
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/610359
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>