[infra] Revive symbolize_stack_trace wrapper script.
This stopped working and was removed in:
https://skia-review.googlesource.com/c/skia/+/538636
due to lack of resources to fix it.

This brings it back and fixes it, requiring a downgrade
to DWARF 4 symbols (gn/skia/BUILD.gn) and some Python 3
changes (encoding='utf-8'). As it did before, this should
stream the output to stdout (tests can take a long time
to run) and then processes the last 500 lines looking for
a stacktrace. This uses the system addr2line binary to
process that output into human-readable file/function/line.

Example output:
https://logs.chromium.org/logs/skia/5ed8b4dff8d56f11/+/steps/symbolized_dm/0/stdout

Change-Id: I12187118c4b09394baea54ac795ebcedf50451e8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/610659
Reviewed-by: Ravi Mistry <rmistry@google.com>