[metal] Fix issues with non-AA clip.
When using non-AA clip we can end up using the stencil attachment.
In this situation we will use the same OpsRenderPass to draw to the
stencil and then to the color attachment. However, this generates
different MTLRenderCommandEncoders. If we start the OpsRenderPass
with a color Clear loadAction, we never reset it to Load for subsequent
MTLRenderCommandEncoders, so we keep clearing the color attachment
over and over.

Bug: skia:13969
Change-Id: I9ba85388ddfed9753c86e57381b6dc761206dad6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/614058
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>