Add RP codegen support for the ?: ternary operator.
This uses copy-masked-slots to combine the results of the true- and
false-expressions, so it doesn't need any new RP ops. It does rely
on the ability to create a second stack, so that the we don't have
to worry about the test-expression taking up space on the primary
stack.

Bug: skia:13676
Change-Id: I580608dceb741ed58e45c6b515052f43851d7214
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/614736
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>